requirements.txt
anthropic>=0.40.0
openpyxl>=3.1.0
jsonschema>=4.20.0
pandas>=2.0.0
streamlit>=1.30.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
